atualizei a tela de confirmação e deixei só uma tela para confirmação de concluído e negado
