+++ App/view/login.py
@@ -1,7 +1,7 @@
 from PyQt5.QtWidgets import QDialog, QMenu
 from PyQt5.QtCore import Qt, QTimer, QPoint
 from PyQt5.QtCore import pyqtSlot
-from .feadbackErro import FeadbackErro
+from .feedback import Feedback
 from PyQt5.uic import loadUi
  
  
@@ -88,7 +88,8 @@ class LoginInterface(QDialog):
  
     def dadosInvalidos(self):
         texto = 'DADOS INCOMPLETOS.'
-        resposta = FeadbackErro(texto)
+        resposta = Feedback
+    (texto)
         # resposta.mudarFoto("Validado")
         if resposta.exec_():
             print('erro')
