+++ App/view/telaConfirmacao.py
@@ -4,10 +4,14 @@ from PyQt5.QtCore import Qt
 
 
 class TelaConfirmacao(QDialog):
-    def __init__(self, aviso, txtBtnOk):
+    def __init__(self, titulo, aviso, txtBtnOk):
         super().__init__()
         loadUi('App/view/ui/telaConfirmacao.ui', self)
-        self.aviso.setText(aviso)
+        if aviso == '':
+            self.aviso.close()
+        else:
+            self.aviso.setText(aviso)
+        self.titulo.setText(titulo)
         self.btnOk.clicked.connect(self.accept)
         self.btnCancelar.clicked.connect(self.reject)
         self.btnFechar.clicked.connect(self.reject)
@@ -15,12 +19,14 @@ class TelaConfirmacao(QDialog):
         # Remove a barra de título e as bordas da janela
         self.setWindowFlags(Qt.FramelessWindowHint)
         
+        self.setAttribute(Qt.WA_TranslucentBackground)
+        
         self.btnOk.setText(txtBtnOk)
         
 if __name__ == "__main__":
     from PyQt5.QtWidgets import QApplication
     app = QApplication([])
-    info = TelaConfirmacao('Teste', "Sim")
+    info = TelaConfirmacao('txt', 'Teste', "Sim")
     if info.exec_():
         print('ok')
     else:
