+++ App/view/ui/feadbackConcluido.ui
@@ -6,8 +6,8 @@
    <rect>
     <x>0</x>
     <y>0</y>
-    <width>834</width>
-    <height>156</height>
+    <width>810</width>
+    <height>163</height>
    </rect>
   </property>
   <property name="windowTitle">
@@ -103,6 +103,9 @@
         </property>
         <property name="alignment">
          <set>Qt::AlignCenter</set>
+        </property>
+        <property name="wordWrap">
+         <bool>true</bool>
         </property>
        </widget>
       </item>
